core/types: Populate Jovian receipt fields (#710)
* core/types: Move receipt tests OP diff to separate file

* core/types: Populate Jovian receipt fields
