+++ core/types/gen_receipt_json.go
@@ -42,6 +42,7 @@ func (r Receipt) MarshalJSON() ([]byte, error) {
 		L1BlobBaseFeeScalar   *hexutil.Uint64 `json:"l1BlobBaseFeeScalar,omitempty"`
 		OperatorFeeScalar     *hexutil.Uint64 `json:"operatorFeeScalar,omitempty"`
 		OperatorFeeConstant   *hexutil.Uint64 `json:"operatorFeeConstant,omitempty"`
+		DAFootprintGasScalar  *hexutil.Uint64 `json:"daFootprintGasScalar,omitempty"`
 	}
 	var enc Receipt
 	enc.Type = hexutil.Uint64(r.Type)
@@ -68,12 +69,9 @@ func (r Receipt) MarshalJSON() ([]byte, error) {
 	enc.FeeScalar = r.FeeScalar
 	enc.L1BaseFeeScalar = (*hexutil.Uint64)(r.L1BaseFeeScalar)
 	enc.L1BlobBaseFeeScalar = (*hexutil.Uint64)(r.L1BlobBaseFeeScalar)
-	if r.OperatorFeeScalar != nil {
-		enc.OperatorFeeScalar = (*hexutil.Uint64)(r.OperatorFeeScalar)
-	}
-	if r.OperatorFeeConstant != nil {
-		enc.OperatorFeeConstant = (*hexutil.Uint64)(r.OperatorFeeConstant)
-	}
+	enc.OperatorFeeScalar = (*hexutil.Uint64)(r.OperatorFeeScalar)
+	enc.OperatorFeeConstant = (*hexutil.Uint64)(r.OperatorFeeConstant)
+	enc.DAFootprintGasScalar = (*hexutil.Uint64)(r.DAFootprintGasScalar)
 	return json.Marshal(&enc)
 }
 
@@ -106,6 +104,7 @@ func (r *Receipt) UnmarshalJSON(input []byte) error {
 		L1BlobBaseFeeScalar   *hexutil.Uint64 `json:"l1BlobBaseFeeScalar,omitempty"`
 		OperatorFeeScalar     *hexutil.Uint64 `json:"operatorFeeScalar,omitempty"`
 		OperatorFeeConstant   *hexutil.Uint64 `json:"operatorFeeConstant,omitempty"`
+		DAFootprintGasScalar  *hexutil.Uint64 `json:"daFootprintGasScalar,omitempty"`
 	}
 	var dec Receipt
 	if err := json.Unmarshal(input, &dec); err != nil {
@@ -193,6 +192,9 @@ func (r *Receipt) UnmarshalJSON(input []byte) error {
 	}
 	if dec.OperatorFeeConstant != nil {
 		r.OperatorFeeConstant = (*uint64)(dec.OperatorFeeConstant)
+	}
+	if dec.DAFootprintGasScalar != nil {
+		r.DAFootprintGasScalar = (*uint64)(dec.DAFootprintGasScalar)
 	}
 	return nil
 }
